requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
auto-py-to-exe==2.42.0
backoff==2.2.1
bcrypt==4.1.2
bottle==0.12.25
bottle-websocket==0.2.9
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.7
cx-Logging==3.1.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
Eel==0.16.0
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
google-auth==2.26.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.0.345
langchain-core==0.0.13
langsmith==0.0.79
lief==0.13.2
looseversion==1.3.0
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.3
pefile==2023.2.7
pillow==10.2.0
posthog==3.3.1
protobuf==4.25.1
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pypdfium2==4.26.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
whichcraft==0.6.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
requests
numpy